Finished handling and testing 'i', 'd', 's', 'c'. Started to handle input 'p'.

+++ printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: tphlogis <tphlogis@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/19 18:43:42 by tphlogis          #+#    #+#             */
-/*   Updated: 2021/05/06 14:43:48 by tphlogis         ###   ########.fr       */
+/*   Updated: 2021/05/09 22:02:21 by tphlogis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,6 +73,6 @@ char			*find_chrs_in_str(const char *s, char *chrs);
 void			*ft_setmem(void *s, int c, size_t len);
 char			*itoa_withno_malloc(char* dst, long n);
 int				digit_nbr(long n);
-unsigned int	get_unsigned(long n);
+unsigned long	get_unsigned(long n);
 long			max(long a, long b);
 #endif
